vsite/requirements.txt
awsebcli==3.14.7
Babel==2.6.0
botocore==1.12.62
cement==2.8.2
colorama==0.3.9
Django==2.1.3
django-phonenumber-field==2.1.0
docutils==0.14
jmespath==0.9.3
pathspec==0.5.5
phonenumbers==8.10.2
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
urllib3==1.22
